requirements.txt
setuptools~=57.0.0
aiogram~=2.17.1
pymongo~=4.0.1
dnspython~=2.1.0
python-dotenv~=0.20.0